asvspoof2021/LA/bio_variable/requirements.txt
absl-py==1.3.0
appdirs==1.4.4
audioread==3.0.0
auditok==0.2.0
brotlipy
cachetools==5.2.1
certifi==2022.12.7
cffi 
charset-normalizer 
cryptography 
cycler==0.11.0
decorator==5.1.1
fonttools==4.38.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.7.0
idna 
importlib-metadata==6.0.0
joblib==1.2.0
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numba==0.56.4
numexpr==2.8.4
numpy==1.21.6
oauthlib==3.2.2
olefile 
packaging==23.0
pandas==1.3.5
Pillow 
pooch==1.6.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser 
pyOpenSSL
pyparsing==3.0.9
PySocks 
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests 
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
samplerate==0.1.0
scikit-learn==1.0.2
scipy==1.7.3
six 
soundfile==0.11.0
spafe==0.1.2
tables==3.7.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions 
urllib3 
Werkzeug==2.2.2
zipp==3.11.0
